+++ src/vue/index.ts
@@ -3,11 +3,17 @@
  *
  * @packageDocumentation
  */
+export * from "./breakpoints";
+export * from "./element-size";
 export * from "./emits";
+export * from "./event-listener";
 export * from "./expose";
 export * from "./func";
 export * from "./install";
+export * from "./now";
 export * from "./props";
 export * from "./render";
+export * from "./resize-observer";
 export * from "./slots";
+export * from "./window-size";
 export * from "./with";